fix bug with cannon speed
Cannon ball speed is increased.
Fixed the bug where balls are slow before the ship has moved.
Normalised the ships diagonal speed to make it consistent with cannon ball speed.
There is a limit somewhere on the max speed of the balls so I've set them to go slightly slower than that.